+++ src/b-tree/index.ts
@@ -3,3 +3,4 @@ export * from './expression';
 export * from './token';
 export * from './negation';
 export * from './filter';
+export * from './operations';
