+++ src/parser.d.ts
@@ -1,4 +1,4 @@
 import { Node } from './b-tree';
 import { v4 as uuid } from 'uuid';
 
-export function parse(str: string): Node | {};
+export function parse<N extends Node | {}>(str: string): N;
